refactoring ConversationListViewController
- extensions to separate files
